Rename custom Jest matchers folder
