+++ test/native/matchers/to-be-visible.js
